00-prompting-techniques/module-01-guidelines/requirements.txt
google-genai

python-dotenv